reduce utils.__init__.py by moving functions in another file
